Fix mariadb cluster role
The mariadb cluster role expected a wrong name for the analytics API
settings variable.

Add support for the notes, insights, ecommerce, and programs databases.

The old "check that this is the first server in the group" strategy
doesn't work with the Ansible 1.9+.  Update it accordinly, with
"run_once".

Make the MARIADB_HAPROXY_USER lines be idempotent, by using
INSERT IGNORE.

Rename common migration password variable to match the default setting
in the edxapp role.